Update to Complete::Getopt::Long 0.471, no longer uses Perinci::Sub::To::CLIDocData.
Perinci::Sub::To::CLIDocData is not appropriate to use here. Instead, we now
pass ggls_res to Complete::Getopt::Long's 'extras'. CGL will use this
information as well as 'cmdline' ('common_opts') and 'r' ('meta') to extract the
summaries. Yes, this is Perinci::CmdLine specifics which leaks into CGL for now.